Update 07. Kadane Algo (subarray sum + subarray return).cpp

+++ EY/07. Kadane Algo (subarray sum + subarray return).cpp
@@ -7,7 +7,7 @@ pair<int, vector<int>> maxSubArraySum(const vector<int>& nums) {
     int maxSum = nums[0], currentSum = nums[0];
     int start = 0, end = 0, tempStart = 0;
 
-    for (size_t i = 1; i < nums.size(); ++i) {
+    for (int i = 1; i < nums.size(); ++i) {
         if (nums[i] > currentSum + nums[i]) {
             currentSum = nums[i];
             tempStart = i;  // Start a new subarray
